requirements.txt
pandas
numpy
matplotlib
biopython
python-dateutil
streamlit
jupyter
